+++ web/src/main/java/com/homeapp/javatraining/service/UserStatisticsServiceImpl.java
@@ -3,6 +3,7 @@ package com.homeapp.javatraining.service;
 import com.homeapp.javatraining.dto.UserTopicStats;
 import com.homeapp.javatraining.model.TestResult;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -10,6 +11,7 @@ import java.util.List;
 import java.util.Map;
 
 @Slf4j
+@Service
 public class UserStatisticsServiceImpl implements UserStatisticsService {
 
     @Override