+++ web/src/main/java/com/homeapp/javatraining/util/ValidationFactory.java
@@ -1,6 +1,5 @@
 package com.homeapp.javatraining.util;
 
-import com.homeapp.javatraining.handler.ErrorHandler;
 import com.homeapp.javatraining.validation.QuestionValidator;
 import com.homeapp.javatraining.validation.UserValidation;
 
@@ -15,9 +14,5 @@ public class ValidationFactory {
 
     public static QuestionValidator createQuestionValidator() {
         return new QuestionValidator();
-    }
-
-    public static ErrorHandler createErrorHandler() {
-        return new ErrorHandler();
     }
 }